requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.34.9
botocore==1.34.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.0
cryptography==41.0.7
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==21.0.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
moto==4.2.12
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
py-partiql-parser==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
responses==0.24.1
rpds-py==0.15.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
xmltodict==0.13.0
